Python_AI/requirements.txt
annotated-types==0.7.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
faiss-cpu==1.10.0
filelock==3.17.0
fsspec==2025.2.0
google-auth==2.38.0
google-genai==1.0.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
setuptools==75.8.0
sympy==1.13.1
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2